Modified to use a JSON API for the list
One drawback to hardcoding the list in the theme is that as new documentation is
added, we need to trigger a new build for every component.

This patch takes a different approach: it fetches the list via XHR, and then
injects entries into the DOM.

Additionally, this patch does a number of fixes to the grid layout. While it's
no longer ordered by column, it displays evenly and responsively.