Switch to getopts for script arguments
This patch switches to getopts for handling script arguments (vs using
environment variables).

Additionally, it notes we will no longer be using travis to build docs;
this is because currently no repos have a valid OAuth2 token associated
with them, due to a security breach several months ago. With the
addition of zfbot to our infrastructure, we can use it for documentation
build automation as well.