Allows building docs from either doc or docs dir
GitHub uses either the `.github/` or `docs/` directories to hold things
such as the contributors guide, code of conduct, issue and pull request
templates, etc. To reduce the number of folders required in the
top-level directory, we can easily re-purpose our own documentation
directory, as we place documentation under its `book/` subdirectory, and
build in the `html/` subdirectory.

This patch adds support to detect if a `docs/` directory exists, and, if
so, it will assume that that directory should be used as the root for
documentation.