Allow "hidden" pages
This is a hack learned from the mkdocs issue tracker:

- https://github.com/mkdocs/mkdocs/issues/699#issuecomment-176948725

Essentially, with this in place, if a page _title_ begins with `_`, it
will not be included in the navigation.

Additionally, this patch goes a step further, and the rendered markdown
of such pages is injected only into the `<head>` section of the built
document, with no other content. This allows for a very quick redirect,
with no content present to index.

This is useful when you want to preserve link continuity. The old page
can continue to exist, but no pages on the site will reference it, which
means that search engines will eventually stop crawling it.

The stated recommendation in the README is to replace the contents of
that page with JS and/or a meta refresh in order to force a redirect
once loaded; most search engines will notice these and honor them,
removing the page from their search results.