Fixes issue with hidden navigation sections
Previously, if a nav item had children, it was always being displayed,
even if it matched the "hidden" rule (starts with `_`), and even if its
own hidden children were not displayed.

This puts the check for the prefix outside the `nav_item.children`
check, which also, conveniently, allows us to move it out of the else
statement when rendering a nav item.