Allows individual components to opt-in to a custom landing page
Modifies the `build.sh` script to look for a file in the package root,
`.zf-mkdoc-theme-landing`; when found, it will run the `swap_index.php`
script.

Alters the `main.html` file of the theme to add the missing comments
that allowed swapping in content.

This approach ensures that components that have the `index.html` defined
**will not** swap in their contents **unless** a commit is made that
adds the `.zf-mkdoc-theme-landing` file.