Ensure _any_ component dropdown is populated.
Previously, we selected the one _not_ in the navbar specifically, but
this meant that those using the navbar could not use the dropdown.

Interestingly, while `Choices.js` allows you to provide a list of
elements... it will not actually populate them, only the _first_ in the
list, which is why I was not seeing them before (it was populating the
one in the navbar, which is first when searching the DOM).

This patch selects _all_ of the elements, casting them to an array; it
then loops through each, creating the Choices.js select dropdown and
populating it.