Ensure the space for the dropdown expands when needed
If the parent container of the `.component-selector` element is short,
then the `Choices.js` dropdown expands to fill only the current height.
This is problematic in the navbar especially, but also in the sidebar
when there's no entries below it (or few entries).

This patch adds listeners when showing and hiding the dropdown that
expand and contract the parent container's height respectively. I've
hard-coded 350px, as that appears to be the height of the component list
once expanded; unfortunately, it's too difficult to determine what the
height is until _after_ it's expanded, at which point the event has
already completed firing.