fix: anchors.js was not working
Due to a problem with how the main `anchor.js` file is written, it
cannot be minified cleanly. As such, I've removed it, and modified the
theme to require it via CDNJS.

Additionally, I've moved the logic that creates the anchors into its own
`<script>` tag in the main template, as otherwise, the `anchors` element
was not in scope. This fixes all issues with anchors.js working.